KDKPAS-2098 Enable metadata scraping using FFMpeg for AVI files.
This was done as per https://gitlab.csc.fi/dpres/file-scraper/merge_requests/16#note_22958 and the accompanying RocketChat discussion. Getting the colour information from MediaInfo is not straightforward, but the FFMpeg metadata scraping already had it implemented. All functionality is not in yet and the tests have not been edited accordingly.