KDKPAS-2098 Fix container detection and add MIME type.S-2098 Fixi container detection and add MIME type.
FFMpeg returns stream and container data separately. Whether a file is
of a type with a container is also known only after the first metadata
model is constructed. To prevent accidentally creating more metadata
models than there's data for (one for each stream + one for a
"container" that doesn't exist), a guard clause was added to prevent
this when a file does not have a container.

Now the simple metadata model also tries to determine the MIME type
based on format name. For some file types the result is ambiguous, so
(:unav) is returned (in those cases other scrapers are needed for
metadata collection in any case so this should not be a problem). The
really metadata-scraping metadata model also checks codec name if
format_name is not available.