KDKPAS-2098 Improve codec quality test material for AVI/JPEG2000.
It was decided that quality is reported as lossless only if neither
wavelet transform nor possible chroma subsampling has reduced the
quality of the stream. The old "lossless" video clip was in yuv420p
format, i.e. chroma subsampling was used even though the wavelet
transform was lossless. This file was renamed and a new one constructed
without subsampling.