KDKPAS-2241 Simpler solution to MIME type conflict with headerless XML.
The work done for ticket KDKPAS-2198 introduced a bug: forced MIME type
and version were no longer passed to metadata models used for XML and
XHTML+XML files. The XML declaration is not mandatory, but files
without it are not currently detected as XML, so scraping XML files
without it as text/xml instead of text/plain requires using file type
forcing. Unfortunately, this bug meant that metadata models used by
MagicScraper did not get the memo, resulting in MIME type being reported
as text/plain, which differs from text/xml from XmllintScraper, causing
the scraper to raise an exception. This bug is caught by dpres-ipt unit
tests but was not covered by scraper tests.

This commit fixes the bug and adds headerless XML test data. Due to XML
files without the XML declaration still not being detected as XML, this
file is not used in normal e2e tests but in filetype-forced ones
instead.