KDKPAS-2241 Make sure dpres-xml-schemas is installed with file-scraper-full
The schemas are required for validating e.g. METS documents. Previously, dpres-xml-schemas was not listed as a dependency for file-scraper-full, possibly causing the file to be classified as not well-formed and metadata not being scraped if the required package is not manually installed. This addition ensures that dpres-xml-schemas is installed alongside file-scraper-full.