KDKPAS-2234 Make CSV tests neater
Now that we have CSV test data stored in files, we can use parse_results
to construct full path of the file based on its name and MIME type. This
makes the code cleaner, more uniform with other scraper tests, and also
more resilient to bugs if we e.g. decide to change the test file path
convention.