KDKPAS-2234 Make MIME type forcing in e3e tests explicit
Files with forced MIME types are no explicitly listed, which makes the
code more resilient to cases such as intorudcing new test file with
string "text_csv" as a part of the path somewhere else than in the MIME
type. This could happen if we for some reason want to e.g. test scraping
an XML file with fields containing both textual and CSV data, a case
which should definitely not be scraped as CSV.