KDKPAS-2375 Print a short message if scraper raises an exception
Scraper usually doesn't crash, but if it does, instead of blurting out the
full stack trace, the CLI now just prints the message from the raised
exception and exits. This message is likely often enough to identify the
problem (e.g. providing a non-existent schematron path), but if it isn't
the user should resort to operating the Python class directly and/or
using a debugger to get a better view of the situation.