KDKPAS-2298 Preliminary list of important/ignored warnings
These files contain possible warnings grepped from v.Nu source, quickly
sorted into two categories: important ones that we want to notice and
others we should ignore. These files are not usable as is, both because
they are just code snippets but also because the separation to
important/ignored is not final.