KDKPAS-2485 Add test file for which vnu produces unicode normalization warning
Currently this file is reported as not well-formed, which can be
confirmed by scraping the file:
```
>>> from file_scraper.scraper import Scraper
>>> s = Scraper("tests/data/text_html/valid_5.0_unicode_normalization_warning.html")
>>> s.scrape()
```
and examining `s.info`, which at the moment has "info warning: Text runi
is not in Unicode Normalization Form" in errors list of `VnuScraper`.

The file is named as "valid" though: the goal of the ticket this commit
relates to is to log this warning as an "info" level message, not as an
error.