TPASPKT-782 Implement ContainerGrader
Implement ContainerGrader for grading video containers based on the
allowed combinations defined in the File Formats specification, section
6, tables 2 and 3.

Since multiple graders are used per single file, select the lowest grade
assigned by any of the graders, as it's likely to be the correct one.

This is the case with ContainerGrader, which inspects individual stream
results. This is in constrast to Grader, which only checks the main MIME
type and version, and is thus more likely to report a more positive
grade.