KDKPAS-2606 Remove well-formedness check from JHove PDF version reporting
We don't know of a case in which it would be important to not read the
version field of a file if it is not well-formed, so the check was
unnecessary. With the check in place, it would also have been possible
that JHove erroneously reports a PDF 1.7 file (that it does not support)
as not well-formed, but we aren't able to discard that result due to
versio being :unav.