KDKPAS-2785 Fix an old bug in JHove test logic
Gif and PDF tests for JHove suffered from a bug that caused only the
first test file in a for loop to be tested. Testing all of the files
revealed some testing errors that were also fixed. Additionally fixed a
docstring in JHove model's `version` function that was not edited when the
changes to that function were made.