KDKPAS-2894 Add test case for file-5.30 recursion bug
Add a JPEG file with enough JPEG segments to break file-5.30, which
uses recursion to read metadata segments and crashes when exceeding the
default recursion depth of 30.

The test file was created by generating a ICC color profile using
DisplayCAL, and attaching both the ICC profile and additional GPS
metadata into the same JPEG file. This is enough to cause file 5.30 to
crash unless the recursion depth is raised. The issue was fixed upstream
in file 5.39.