TPASPKT-1243 Put VeraPDF warnings to messages
VeraPDF warnings are printed to stderr, causing file-scraper to
interpret the file as not well-formed, even if there are no severe
errors but only warnings.

On VeraPDF detector omit the warnings by setting the log level when
calling VeraPDF. Warnings are not useful in the detection phase.

When scraping, the warnings are still useful information, so dump stderr
to messages when we can be entirely sure that the file is valid and
stderr contains no severe errors.