TPASPKT-1278 Better tests for encoded filenames
Currently, the Scraper object encodes provided filename to Bytes object
using the file system encoding. The encoded filename is then passed to
scrapers and detectors. However, all unit tests of scrapers and
detectors use string filenames instead of bytes filenames. This leads to
a situation where unit tests might pass but the e2e-test fails, because
the scraper/detector works when filename is passed as string, but it
does not work when filename is passed as bytes. And debugging these
errors might not be easy.

In this commit, filename encoding is added to all scrapers.

I also added a test for scraping file with undecodable filename,
because undecodable filenames are the reason for storing the filename as
bytes, althought it does not currently work.