TPASPKT-1382 Use Java 8 for dbtpk if available
Java 17 is installed and set by default on RHEL9, but it's not
compatible with dbptk, which uses Java 8 (aka 1.8.0). Set `PATH`
environment variable so that Java 8 is used if available; otherwise,
fall back to whatever Java interpreter `java` resolves to.

The specific Java interpreter is installed as a dependency via
`dbptk-developer`, so there's no need to set the dependency explicitly.