KDKPAS-3605 Update invalid mp3 test files
To match the file name with the FFMpeg error ("Header missing"), renamed invalid_1_missing_data.mp3 to invalid_1_missing_header.mp3.

Removed invalid_1_wrong_version.mp3, because it was only invalid due to a trailing newline, which was fixed.

Created a new invalid_1_missing_data.mp3, which is identical to valid_1.mp3, but is missing the bytes from 0x3630 to 0x383F. This file creates the FFMpeg error "invalid new backstep".