KDKPAS-3594 Pass a list of used file types to PIL
Our users have TIFF files which PIL erroneously detects as FLI/FLIC.
Pass a list of file types to use to PIL in order to skip FLI/FLIC
detection. The file type list was gathered by debugging the PIL code.
Due to it being constructed while importing plugins in PIL, it is not
easily programmatically available.