TPASPKT-1418 Improve filetype updating.
Separate updating mimetypes and the version.

Move check_wellformed from a parameter to a property for more consistency inside scraper.py.

Change UnknownValue.UNAV back to UNAV because replacement of the unknown values is wise to do int the defaults of scraper.

Fix Bugs related to version and mimetype assignment.

Add more logging.