TPASPKT-1418 Remove PredefinedDetector
If the detector results are suppose to be compared to the user given input it doesn't make sense that one of the detectors returns user defined input as an output.

PredefinedDetector needs to be deleted or moved out from the list of detectors. So the results it provides doesn't interfere with the comparison to the user output.