KDKPAS-3663 Scrape with Wand in separate process
Wand is currently suspectible to memory leaks when multiple image files
are scraped in one Python process. The memory growth is not
deterministic and seems related to unreliable Python garbage collection,
with `wand.image.Image` not always being deleted by Python immediately
in response to zero reference count and sometimes also evading deletion
by Python's secondary garbage collector (see Python module `gc` for
details).

While `gc.collect()` seems to help in some instances, it is not entirely
reliable. Explicitly calling `destroy()` on supported objects can also
alleviate the issue but not entirely prevent it.

Furthermore, installing newer versions of either Wand or ImageMagick6
(the underlying library wrapped by Wand) does not seem to help either.
The issue was also reproduced in the newest upstream version available
in nixpkgs' unstable channel, indicating that it's probably an upstream
issue and not related to RHEL9/Alma patches.

As a workaround, perform each Wand scraping in a separate Python process
and marshal the result over as `namedtuple` instances. This ensures that
any objects missed by the garbage collector are deleted at the same time
as the single-use scraping process.

Co-authored-by: Mikko Vatanen <mikko.vatanen@csc.fi>
Co-authored-by: Juho Kuisma <juho.kuisma@csc.fi>