TPASPKT-1539 Add UserdefinedMeta
The users given metadata is now initialized when the Scraper gets created.

The _merge_results needs to generate the result stream dict twice, because otherwise errors wouldn't be generated from the top level function.
This is not optimal, but works for now.

Another option would've been to overwrite the metadata values as
userdefined at the end, but it would've been a patch as well to a
systematic problem.