TPASPKT-1617 Fix subrip test-file.
For some reason the current subrip test-file won't be detected as a subrip by the magic detector normally. The reason for this is the numbering starting from 0 instead of 1 even though the various scattered internet sources and example files don't specify the starting number. This is probably a bug or choice from the file-magic developers to start the numbers from 1 instead of 0 or some other arbitrary number. It seems to be a convention instead of a hard rule in the specification.

But for our purposes it is enough that the magic detector actually detects the mime type.