TPASPKT-1617 Improve TextEncodingExtractor
The character set shouldn't be checked and invalidaded in the extraction part of the process because that causes the file to be deemed not well formed based on external characteristics. It is better to fail the file by reading it and failing to decode it instead.

The extractors is_supported function should be used instead to determine if the extractor will be used or not.