TPASPKT-1615 Move common Exif property to base
`stream_type` is common to both JPEG/Exif and DNG
files; move them to the base class.

`byte_order` was not moved although it initially appeared common to both
of them. JHOVE reported little-endian for the same file, while ExifTool
reported big-endian, though these results could refer to different parts
of the file and not the file as a whole.