TPASPKT-1614 Add higher level function to Extractors
Add a function to Extractors which makes sure that the extractor results will be validated after each extraction on the base class level.

The _extract function is now the function which will be overwritten in each Extractor and the higher order function extract shouldn't be overwritten.

Each Extractor also has the _validate function called automatically by the higher order extract function.

Each of the functions can still be overwritten or intercepted if additional functionality separation is necessary.