TPASPKT-1614 Remove user input from results
Collect the user input into a separate variable instead of collecting it into the results for less confusion.

This change prompts a possibility for the removal of at least some predefined values by using a MetadataModel instead of multiple variables.

This change plans for the removal of calling the generate_metadata_dict twice, which is an ugly hack.